requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.5
crewai>=0.1.0
crewai-tools>=0.0.1
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0
chardet>=5.0.0